Move `WidgetRegistry` to *base.py*
