+++ src/core/src/toga/app.py
@@ -10,7 +10,7 @@ from toga.command import CommandSet
 from toga.handlers import wrapped_handler
 from toga.icons import Icon
 from toga.platform import get_platform_factory
-from toga.widget_registry import WidgetRegistry
+from toga.widgets.base import WidgetRegistry
 from toga.window import Window
 
 try:
