+++ src/core/tests/test_app.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock, patch
 
 import toga
 import toga_dummy
-from toga.widget_registry import WidgetRegistry
+from toga.widgets.base import WidgetRegistry
 from toga_dummy.utils import TestCase
